chore: examples
