+++ pkg/function/examples/evenodd/.gitignore
