+++ pkg/function/examples/evenodd/Dockerfile
