+++ pkg/function/examples/flatmap/Dockerfile
