+++ pkg/function/examples/flatmap/Makefile
