+++ pkg/function/examples/flatmap/README.md
