+++ pkg/function/examples/sum/main.go
